Refactoring, add WIP undo
