+++ package.json
@@ -2,7 +2,7 @@
   "name": "prox2",
   "version": "1.0.0",
   "scripts": {
-    "dev": "next",
+    "dev": "env NODE_OPTIONS=--openssl-legacy-provider next",
     "build": "next build",
     "start": "next start",
     "type-check": "tsc"
@@ -11,7 +11,7 @@
     "@slack/web-api": "6.2.4",
     "body-parser": "1.19.0",
     "emojis-keywords": "2.0.0",
-    "next": "11.1.3",
+    "next": "^11.1.3",
     "pg": "8.6.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
@@ -20,7 +20,7 @@
     "typeorm": "0.2.34"
   },
   "devDependencies": {
-    "@babel/core": "7.14.6",
+    "@babel/core": "7.18.6",
     "@babel/plugin-proposal-class-properties": "7.14.5",
     "@babel/plugin-proposal-decorators": "7.14.5",
     "@types/body-parser": "1.19.1",
@@ -30,7 +30,7 @@
     "@types/source-map-support": "0.5.4",
     "airtable": "0.11.1",
     "babel-plugin-transform-typescript-metadata": "0.3.2",
-    "typescript": "4.3.5"
+    "typescript": "^5.2.2"
   },
   "license": "AGPL-3.0-or-later"
 }
